requirements.txt
# Core Requirements
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
scipy==1.11.1

# Web Application
flask==2.3.2
jinja2==3.1.2

# Data Processing
python-dateutil==2.8.2
pytz==2023.3

# Configuration
pyyaml==6.0

# Development
jupyter==1.0.0
ipython==8.14.0
matplotlib==3.7.1
seaborn==0.12.2

# Testing
pytest==7.4.0